Atualiza algoritmo em DeRomano
Atualiza o algoritmo de conversão em DeRomano::paraLong.
Adiciona novos caracteres ao mapa em DeRomano.hpp.
Atualiza instruções de compilação.

+++ conversores/ParaRomano.cpp
@@ -4,7 +4,7 @@
 
 std::string ParaRomano::DeLong(unsigned long numToConvert){
 	std::string resultado = "";
-	for (int i = 0; i < std::size(ParaRomano::VALORES); i++) {
+	for (size_t i = 0; i < std::size(ParaRomano::VALORES); i++) {
 		while (numToConvert >= ParaRomano::VALORES[i]) {
 			numToConvert = numToConvert - ParaRomano::VALORES[i];
 			resultado = resultado + std::string(ParaRomano::SIMBOLOS[i]);
